test: move embedded relation id column tests to functional suite
