+++ test/functional/embedded/embedded-relation-id-column/entity/Avatar.ts
@@ -1,4 +1,4 @@
-import { Entity, PrimaryGeneratedColumn } from "../../../../src"
+import { Entity, PrimaryGeneratedColumn } from "../../../../../src"
 
 @Entity()
 export class Avatar {
